pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gnn-dr"
version = "0.1.0"
description = "GNN-based Dimensionality Reduction via Graph Neural Networks"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.4.0",
    "torch-scatter",
    "torch-sparse",
    "torch-cluster",
    "numpy>=1.24",
    "scipy",
    "scikit-learn",
    "matplotlib",
    "pandas",
    "pytorch-lightning>=2.0",
    "pyyaml",
    "tqdm",
    "transformers",
    "torchvision",
    "wandb",
    "umap-learn",
    "networkx",
]

[project.optional-dependencies]
metrics = ["tensorflow-projection-qm"]
pumap = ["parametric-umap"]
dev = ["pytest", "black", "flake8"]
all = ["gnn-dr[metrics,pumap,dev]"]

[tool.setuptools.packages.find]
include = ["gnn_dr*"]
